requirements.txt
fastapi[all]
pdfminer
supabase

PyMuPDF
thefuzz
requests
PyPDF2
python-pptx